Factor our table module
